splitted cache entry in interface and implementation
