+++ core/src/test/java/com/orientechnologies/orient/core/index/hashindex/local/cache/AbstractLRUListTestTemplate.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 
 import com.orientechnologies.common.directmemory.OByteBufferPool;
+import com.orientechnologies.orient.core.storage.cache.OCacheEntryImpl;
 import com.orientechnologies.orient.core.storage.cache.local.twoq.LRUList;
 import com.orientechnologies.orient.core.storage.cache.OCacheEntry;
 import com.orientechnologies.orient.core.storage.cache.OCachePointer;
@@ -26,12 +27,12 @@ public abstract class AbstractLRUListTestTemplate {
     final ByteBuffer buffer = bufferPool.acquireDirect(true);
 
     OCachePointer cachePointer = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-    lruList.putToMRU(new OCacheEntry(1, 10, cachePointer, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 10, cachePointer, false));
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     Assert.assertTrue(entryIterator.hasNext());
 
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 10, cachePointer, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 10, cachePointer, false));
   }
 
   @Test
@@ -44,16 +45,16 @@ public abstract class AbstractLRUListTestTemplate {
     OCachePointer cachePointerOne = new OCachePointer(bufferOne, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
     OCachePointer cachePointerTwo = new OCachePointer(bufferTwo, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
 
-    lruList.putToMRU(new OCacheEntry(1, 10, cachePointerOne, false));
-    lruList.putToMRU(new OCacheEntry(1, 20, cachePointerTwo, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 10, cachePointerOne, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 20, cachePointerTwo, false));
 
     Assert.assertEquals(lruList.size(), 2);
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     Assert.assertTrue(entryIterator.hasNext());
 
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 20, cachePointerTwo, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 10, cachePointerOne, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 20, cachePointerTwo, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 10, cachePointerOne, false));
   }
 
   @Test
@@ -68,18 +69,18 @@ public abstract class AbstractLRUListTestTemplate {
     OCachePointer cachePointerTwo = new OCachePointer(bufferTwo, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
     OCachePointer cachePointerThree = new OCachePointer(bufferThree, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
 
-    lruList.putToMRU(new OCacheEntry(1, 10, cachePointerOne, false));
-    lruList.putToMRU(new OCacheEntry(1, 20, cachePointerTwo, false));
-    lruList.putToMRU(new OCacheEntry(3, 30, cachePointerThree, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 10, cachePointerOne, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 20, cachePointerTwo, false));
+    lruList.putToMRU(new OCacheEntryImpl(3, 30, cachePointerThree, false));
 
     Assert.assertEquals(lruList.size(), 3);
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     Assert.assertTrue(entryIterator.hasNext());
 
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(3, 30, cachePointerThree, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 20, cachePointerTwo, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 10, cachePointerOne, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(3, 30, cachePointerThree, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 20, cachePointerTwo, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 10, cachePointerOne, false));
   }
 
   @Test
@@ -94,20 +95,20 @@ public abstract class AbstractLRUListTestTemplate {
     OCachePointer cachePointerTwo = new OCachePointer(bufferTwo, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
     OCachePointer cachePointerThree = new OCachePointer(bufferThree, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
 
-    lruList.putToMRU(new OCacheEntry(1, 10, cachePointerOne, false));
-    lruList.putToMRU(new OCacheEntry(1, 20, cachePointerTwo, false));
-    lruList.putToMRU(new OCacheEntry(3, 30, cachePointerThree, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 10, cachePointerOne, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 20, cachePointerTwo, false));
+    lruList.putToMRU(new OCacheEntryImpl(3, 30, cachePointerThree, false));
 
-    lruList.putToMRU(new OCacheEntry(1, 20, cachePointerTwo, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 20, cachePointerTwo, false));
 
     Assert.assertEquals(lruList.size(), 3);
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     Assert.assertTrue(entryIterator.hasNext());
 
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 20, cachePointerTwo, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(3, 30, cachePointerThree, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 10, cachePointerOne, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 20, cachePointerTwo, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(3, 30, cachePointerThree, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 10, cachePointerOne, false));
   }
 
   @Test
@@ -124,20 +125,20 @@ public abstract class AbstractLRUListTestTemplate {
     OCachePointer cachePointerThree = new OCachePointer(bufferThree, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
     OCachePointer cachePointerFour = new OCachePointer(bufferFour, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
 
-    lruList.putToMRU(new OCacheEntry(1, 10, cachePointerOne, false));
-    lruList.putToMRU(new OCacheEntry(1, 20, cachePointerTwo, false));
-    lruList.putToMRU(new OCacheEntry(3, 30, cachePointerThree, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 10, cachePointerOne, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 20, cachePointerTwo, false));
+    lruList.putToMRU(new OCacheEntryImpl(3, 30, cachePointerThree, false));
 
-    lruList.putToMRU(new OCacheEntry(1, 20, cachePointerFour, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 20, cachePointerFour, false));
 
     Assert.assertEquals(lruList.size(), 3);
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     Assert.assertTrue(entryIterator.hasNext());
 
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 20, cachePointerFour, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(3, 30, cachePointerThree, false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 10, cachePointerOne, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 20, cachePointerFour, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(3, 30, cachePointerThree, false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 10, cachePointerOne, false));
   }
 
   @Test
@@ -150,24 +151,24 @@ public abstract class AbstractLRUListTestTemplate {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
 
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
-    lruList.putToMRU(new OCacheEntry(1, 50, cachePointers[5], false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 50, cachePointers[5], false));
 
     Assert.assertEquals(lruList.size(), 11);
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
 
     Assert.assertTrue(entryIterator.hasNext());
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 50, cachePointers[5], false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 50, cachePointers[5], false));
 
     for (int i = 10; i >= 0; i--) {
       if (i == 5)
         continue;
 
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -177,7 +178,7 @@ public abstract class AbstractLRUListTestTemplate {
     ByteBuffer buffer = bufferPool.acquireDirect(true);
 
     OCachePointer cachePointerOne = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-    lruList.putToMRU(new OCacheEntry(1, 10, cachePointerOne, false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 10, cachePointerOne, false));
     lruList.removeLRU();
 
     Assert.assertEquals(lruList.size(), 0);
@@ -192,7 +193,7 @@ public abstract class AbstractLRUListTestTemplate {
     ByteBuffer buffer = bufferPool.acquireDirect(true);
 
     OCachePointer cachePointerOne = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-    OCacheEntry cacheEntry = new OCacheEntry(1, 10, cachePointerOne, false);
+    OCacheEntry cacheEntry = new OCacheEntryImpl(1, 10, cachePointerOne, false);
     lruList.putToMRU(cacheEntry);
     cacheEntry.incrementUsages();
 
@@ -211,7 +212,7 @@ public abstract class AbstractLRUListTestTemplate {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
 
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
     lruList.removeLRU();
@@ -222,7 +223,7 @@ public abstract class AbstractLRUListTestTemplate {
 
     for (int i = 10; i > 0; i--) {
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -235,10 +236,10 @@ public abstract class AbstractLRUListTestTemplate {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
 
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
-    Assert.assertEquals(lruList.remove(1, 50), new OCacheEntry(1, 50, cachePointers[5], false));
+    Assert.assertEquals(lruList.remove(1, 50), new OCacheEntryImpl(1, 50, cachePointers[5], false));
     Assert.assertNull(lruList.remove(1, 500));
 
     Assert.assertEquals(lruList.size(), 10);
@@ -249,7 +250,7 @@ public abstract class AbstractLRUListTestTemplate {
         continue;
 
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -263,11 +264,11 @@ public abstract class AbstractLRUListTestTemplate {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
 
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
     Assert.assertTrue(lruList.contains(1, 50));
-    Assert.assertEquals(lruList.get(1, 50), new OCacheEntry(1, 50, cachePointers[5], false));
+    Assert.assertEquals(lruList.get(1, 50), new OCacheEntryImpl(1, 50, cachePointers[5], false));
 
     Assert.assertFalse(lruList.contains(2, 50));
 
@@ -276,7 +277,7 @@ public abstract class AbstractLRUListTestTemplate {
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     for (int i = 10; i >= 0; i--) {
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -288,7 +289,7 @@ public abstract class AbstractLRUListTestTemplate {
     for (int i = 0; i < 9128; i++) {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
     Assert.assertEquals(lruList.size(), 9128);
@@ -296,7 +297,7 @@ public abstract class AbstractLRUListTestTemplate {
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     for (int i = 9127; i >= 0; i--) {
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -308,18 +309,18 @@ public abstract class AbstractLRUListTestTemplate {
     for (int i = 0; i < 9128; i++) {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
     Assert.assertEquals(lruList.size(), 9128);
 
     for (int i = 0; i < 9128; i++)
-      Assert.assertEquals(lruList.get(1, i * 10), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(lruList.get(1, i * 10), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     for (int i = 9127; i >= 0; i--) {
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -331,18 +332,18 @@ public abstract class AbstractLRUListTestTemplate {
     for (int i = 0; i < 9128; i++) {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
     for (int i = 4564; i < 9128; i++)
-      Assert.assertEquals(lruList.remove(1, i * 10), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(lruList.remove(1, i * 10), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
 
     Assert.assertEquals(lruList.size(), 4564);
 
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
     for (int i = 4563; i >= 0; i--) {
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -355,25 +356,25 @@ public abstract class AbstractLRUListTestTemplate {
     for (int i = 0; i < 9128; i++) {
       ByteBuffer buffer = bufferPool.acquireDirect(true);
       cachePointers[i] = new OCachePointer(buffer, bufferPool, new OLogSequenceNumber(0, 0), 0, 0);
-      lruList.putToMRU(new OCacheEntry(1, i * 10, cachePointers[i], false));
+      lruList.putToMRU(new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
 
-    lruList.putToMRU(new OCacheEntry(1, 0, cachePointers[0], false));
-    lruList.putToMRU(new OCacheEntry(1, 4500 * 10, cachePointers[4500], false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 0, cachePointers[0], false));
+    lruList.putToMRU(new OCacheEntryImpl(1, 4500 * 10, cachePointers[4500], false));
 
     Assert.assertEquals(lruList.size(), 9128);
     Iterator<OCacheEntry> entryIterator = lruList.iterator();
 
     Assert.assertTrue(entryIterator.hasNext());
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 4500 * 10, cachePointers[4500], false));
-    Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, 0, cachePointers[0], false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 4500 * 10, cachePointers[4500], false));
+    Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, 0, cachePointers[0], false));
 
     for (int i = 9127; i >= 1; i--) {
       if (i == 4500)
         continue;
 
       Assert.assertTrue(entryIterator.hasNext());
-      Assert.assertEquals(entryIterator.next(), new OCacheEntry(1, i * 10, cachePointers[i], false));
+      Assert.assertEquals(entryIterator.next(), new OCacheEntryImpl(1, i * 10, cachePointers[i], false));
     }
   }
 
@@ -382,7 +383,7 @@ public abstract class AbstractLRUListTestTemplate {
     final ArrayList<OCacheEntry> entries = new ArrayList<OCacheEntry>();
 
     for (int i = 0; i < 10; i++) {
-      final OCacheEntry cacheEntry = new OCacheEntry(1, i, null, false);
+      final OCacheEntry cacheEntry = new OCacheEntryImpl(1, i, null, false);
 
       entries.add(cacheEntry);
       lruList.putToMRU(cacheEntry);
